+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Master
 
+* Enterprise GitHub support - dbgrandi
 * Use branches for comparison, not commits - orta
 * Breaking: DSL change `files_*` to `*_files` for readability - jeroenvisser101